+++ module-audio/Audio/Profiles/ProfileRecordingOnBoardMic.hpp
@@ -16,8 +16,8 @@
 
 class ProfileRecordingOnBoardMic : public Profile {
 public:
-    ProfileRecordingOnBoardMic(std::function<int32_t()> callback, int8_t gain) : Profile(
-            "Recording On Board Microphone",Type::RecordingBuiltInMic, 0, gain, 0, 0, bsp::AudioDevice::Type::Audiocodec, callback) {}
+    ProfileRecordingOnBoardMic(std::function<int32_t()> callback, float gain) : Profile(
+            "Recording On Board Microphone",Type::RecordingBuiltInMic, 0.0, gain, 0, 0, bsp::AudioDevice::Type::Audiocodec, callback) {}
 };
 
 #endif //PUREPHONE_PROFILERECORDINGONBOARDMIC_HPP
